Add presenter
