+++ src/Contracts/TimeInterface.php
@@ -38,6 +38,7 @@ interface TimeInterface
     public function roundDown(int $precision = 5, bool $roundSeconds = false): TimeInterface;
 
     // Presentation
-    public function present(): TimePresenterInterface;
+    public function toNumericalTime(bool $showSeconds = false): string;
+    public function toReadableTime(bool $showSeconds = false): string;
     public function toString(): string;
 }
