--- src/Contracts/TimePresenterInterface.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace Vdhicts\Time\Contracts;
-
-interface TimePresenterInterface
-{
-    public function asNumericalTime(bool $showSeconds = false): string;
-    public function asReadableTime(): string;
-}
